bsp: enable kmscube
Temporary solution to fix kmscube build, gl3ext file was added.
On salvator-x 8G there is an issue, when starting kmscube
tries to start  channel with connection id for vga,
because it is the first connection id encountered. And the result
is black screen on the screen with HDMI connection.
Fixed the issued by adding command line parameter to kmscube
to enable manual channel id selection

Signed-off-by: Valerii Chubar <valerii_chubar@epam.com>